+++ packages/sdk-core/src/Operation.ts
@@ -18,12 +18,18 @@ export default class Operation {
     readonly populateTransactionPromise: Promise<ethers.PopulatedTransaction>;
     readonly type: OperationType;
 
+    // @note This property is used to ensure BatchCall operations still function
+    // when using the agreement forwarder
+    readonly forwarderPopulatedPromise?: Promise<ethers.PopulatedTransaction>;
+
     constructor(
         txn: Promise<ethers.PopulatedTransaction>,
-        type: OperationType
+        type: OperationType,
+        forwarderPopulatedPromise?: Promise<ethers.PopulatedTransaction>
     ) {
         this.populateTransactionPromise = txn;
         this.type = type;
+        this.forwarderPopulatedPromise = forwarderPopulatedPromise;
     }
 
     /**
@@ -44,13 +50,16 @@ export default class Operation {
     /**
      * Get the populated transaction by awaiting `populateTransactionPromise`.
      * @description Note that we need to populate the txn with the signer.
+     * NOTE: we use the forwarder populated promise if this exists
      * @returns {Promise<TransactionRequest>}
      */
     getPopulatedTransactionRequest = async (
         signer: ethers.Signer
     ): Promise<TransactionRequest> => {
-        const prePopulated = await this.populateTransactionPromise;
-        return await signer.populateTransaction(prePopulated);
+        const txnToPopulate = this.forwarderPopulatedPromise
+            ? await this.forwarderPopulatedPromise
+            : await this.populateTransactionPromise;
+        return await signer.populateTransaction(txnToPopulate);
     };
     /**
      * Signs the populated transaction via the provided signer (what you intend on sending to the network).
