+++ packages/sdk-core/src/interfaces.ts
@@ -44,8 +44,12 @@ export interface ISuperTokenRequestFilter {
 // A better thought out inheritance pattern - SuperToken is parent
 // CFA/IDA inherits and tacks on superToken property
 
+export interface IShouldUseCallAgreement {
+    readonly shouldUseCallAgreement?: boolean;
+}
+
 // write request interfaces
-export interface ISuperTokenModifyFlowParams {
+export interface ISuperTokenModifyFlowParams extends IShouldUseCallAgreement {
     readonly flowRate?: string;
     readonly receiver: string;
     readonly sender?: string;
@@ -141,25 +145,32 @@ export interface IDeleteFlowParams extends IModifyFlowParams {
 }
 
 export interface ISuperTokenUpdateFlowOperatorPermissionsParams {
+    readonly sender?: string;
     readonly flowOperator: string;
     readonly permissions: number;
     readonly flowRateAllowance: string;
+    readonly shouldUseCallAgreement?: boolean;
     readonly userData?: string;
     readonly overrides?: Overrides & { from?: string | Promise<string> };
 }
 
 export interface ISuperTokenFullControlParams {
+    readonly sender?: string;
     readonly flowOperator: string;
+    readonly shouldUseCallAgreement?: boolean;
     readonly userData?: string;
     readonly overrides?: Overrides & { from?: string | Promise<string> };
 }
 
 export interface IUpdateFlowOperatorPermissionsParams
-    extends ISuperTokenUpdateFlowOperatorPermissionsParams {
+    extends ISuperTokenUpdateFlowOperatorPermissionsParams,
+        IShouldUseCallAgreement {
     readonly superToken: string;
 }
 
-export interface IFullControlParams extends ISuperTokenFullControlParams {
+export interface IFullControlParams
+    extends ISuperTokenFullControlParams,
+        IShouldUseCallAgreement {
     readonly superToken: string;
 }
 
@@ -444,6 +455,7 @@ export interface IConfig {
     readonly cfaV1Address: string;
     readonly idaV1Address: string;
     readonly governanceAddress: string;
+    readonly cfaV1ForwarderAddress: string;
 }
 
 export interface IContracts {
